flush and mark before output the end of stream mark; sometimes it still 6 elements missing for q7

+++ pkg/producer_consumer/producer.go
@@ -32,6 +32,6 @@ type MeteredProducerIntr interface {
 	StartWarmup()
 	// GetEventTimeLatency() []int
 	GetCount() uint64
-	NumCtrlMsg() uint64
+	NumCtrlMsg() uint32
 	PrintRemainingStats()
 }
